Fix TSLint warnings in msbot-generic.ts
The following TSLint issues have been fixed:

- typedef
- no-any
- interface-name
- quotemark
- no-empty
- prefer-const
- one-line
- curly
- triple-equals
- newline-before-return
- eofline

A key had to be added to `IConnectGenericArgs` to be able to iterate through its properties. The type of the key had to be `any` to avoid conflict with `for (const key of args.keys)`. An exception for `no-any` has been added to that particular line.